ignore extreme outliers while updating cached int info
say someone accidentally logs -1 billion to a table (because integer
overflow or other reasons). how to detect and prevent these values from
mucking up our cached int stats?

in sybil, if a value comes in that is more than 1000x the SD away from
the current value, we will not update the Min/Max seen value (which are
used to build histograms) because this current value looks quite
spurious

i'm hoping this doesn't lead to too many problems down the road with
bi-modal distributions