add support for reading gzip compressed .db files
sybil can now read (but not write) any .db file that is gzipped and has
'.db.gz' extension. this allows one to selectively gzip any files in the
database. Using gzip decreases the .db file sizes with a small penalty
for read performance. because sybil only ever writes new files, it
should* be safe to gzip any .db file in place

NOTE: this commit also makes sure that the 'rebuild' command supports
gzipped files when rebuilding the table info maps (in case of a
corrupted or lost info.db file).

TEST PLAN:

* run gzip the main table info.db. verify queries still work
* gzip a block info.db, verify queries are accurate
* gzip all .db files in a table. verify queries still work
* remove main table info.db file with all files gzipped. run rebuild
command and verify columns are restored

USE CASE:

two example use cases are: automatically gzipping any data that falls
out of a time window or gzipping everything all the time.

a simple example of zipping all files that are not info or ingest logs is:

```
    for f in `find ./db/TABLE/ -name "*.db" | grep -v info | grep -v ingest_`; do
      gzip ${f};
    done
```

this command is undoable with:

```
    for f in `find ./db/TABLE/ -name "*.db.gz"`; do
      gunzip ${f};
    done
```