[query cache] cache queries via a per block query cache
This commit adds a per block query cache based on the contents of the
querySpec. The query cache is enabled for table, time series and
distribution queries.

To cache a querySpec, its QueryDetails are hashed into an md5sum and
checked for on disk. If they exist, we can re-use those results for the
whole block, saving us some computation time.

To enable the query cache, use the -cache-queries flag. The first query
will be slightly slower because we write to disk, but the subsequent
queries will be faster.

Test Plan: Run a query with and without -cache-queries, plug in to
running snorkel instance and watch dashboards load more quickly.

Tests Added: see TESTPLAN.md