[compression] adjust CARDINALITY_THRESHOLD for columnar compression
After doing some testing, lower the cardinality threshold to 5000 unique
values. This can likely be lowered even further, but that remains to be
seen.

The following are block sizes for various amounts of timestamps stored
in one block, from 0.25 hour spread to 24 hour spread. All of them are
stored in the "high cardinality mode". When stored in low cardinality
mode, between 75 - 100K extra space is used.

-rw-rw-r-- 1 okay okay 167K Feb  5 06:33 tsdb/test_0.25/block803942162/int_time.db
-rw-rw-r-- 1 okay okay 180K Feb  5 06:33 tsdb/test_0.5/block998479838/int_time.db
-rw-rw-r-- 1 okay okay 186K Feb  5 06:33 tsdb/test_1/block350848165/int_time.db
-rw-rw-r-- 1 okay okay 190K Feb  5 06:33 tsdb/test_3/block090475483/int_time.db
-rw-rw-r-- 1 okay okay 192K Feb  5 06:33 tsdb/test_6/block741306497/int_time.db
-rw-rw-r-- 1 okay okay 196K Feb  5 06:33 tsdb/test_12/block050239454/int_time.db
-rw-rw-r-- 1 okay okay 217K Feb  5 06:33 tsdb/test_24/block212806080/int_time.db
-rw-rw-r-- 1 okay okay 235K Feb  5 06:33 tsdb/test_48/block571945678/int_time.db