[SavedRecordBlock] KeyTable should be stored with each ingestion batch
This diff saves ingested records as type SavedRecordBlock.  A
SavedRecordBlock consists of a list of Records and a KeyTable that was
used to generate them.  If there is no KeyTable available in the SRB,
sybil will use the table's current KeyTable during ingestion.

The feature is gated behind -save-srb flag, for now

This addresses 99 and 105