[SavedRecordBlock] KeyTable should be stored with each ingestion batch
This diff saves ingested records as type SavedRecordBlock. This new
method of storing ingestion files is still able to read old ingestion
files. New ingestion files generated by sybil will not be readable by
old sybil binaries.

A SavedRecordBlock consists of a list of Records and a KeyTable that was
used to generate them.  If there is no KeyTable available in the SRB,
sybil will use the table's current KeyTable during ingestion.

This addresses 99 and 105