[block info] only do an lstat on a block's dir if we need to read the files
This makes LoadRecords(nil) faster when we aren't going to print the
table info. Specifically, this will speed up the digest path as
a table grows larger and larger.

For 2100 blocks on my 130mm record jaeger test table, a second is spent
doing unecessary lstats. This code path scales linearly in time, so
twice as many blocks will take twice as long to load.